Add uploader component to screenshots page
