+++ src/routes/screenshots/Batch.svelte
@@ -1,9 +1,12 @@
 <script>
   import { invoke } from "@tauri-apps/api/tauri";
   import { batchPath, isProcessingBatch } from "./store.js";
+  import Uploader from "../../components/uploader/Uploader.svelte";
+  import { parseFile } from "../../components/uploader/uploader.js";
 
   let batches = [];
   let current_batch = null;
+  let files = [];
 
   $: !$isProcessingBatch && listBatches();
 
@@ -11,6 +14,19 @@
     batches = await invoke("list_batches", {
       baseBatchPath: $batchPath,
     });
+  }
+  async function triggerUpload(batch) {
+    let res = [];
+    for (let screenshot of batch.items) {
+      let dataUri = await invoke("get_screenshot_uri", {
+        screenshot: screenshot,
+      });
+      let name = /MapleLegends[_ ](.*).png/.exec(screenshot.name)[0];
+      let item = await parseFile(name, dataUri);
+      res.push(item);
+      console.log(item);
+    }
+    files = [...res];
   }
 </script>
 
@@ -22,10 +38,11 @@
         <a
           href={"javascript:void(0)"}
           on:click={() => {
-            current_batch = current_batch == batch.name ? null : batch.name;
+            current_batch = current_batch == batch ? null : batch;
           }}>{batch.datetime}</a
         >
-        {#if current_batch == batch.name}
+        <button on:click={() => triggerUpload(batch)}>Upload to OwlRepo</button>
+        {#if current_batch && current_batch.name == batch.name}
           <ul>
             {#each batch.items as item}
               <li>{item.datetime}</li>
@@ -35,4 +52,8 @@
       </li>
     {/each}
   </ul>
+{/if}
+
+{#if files.length > 0}
+  <Uploader {files} />
 {/if}
